requirements.txt
fastapi==0.85.0
sqlalchemy==1.4.39
uvicorn==0.18.2
pydantic==1.10.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv
python-multipart
alembic==1.7.7
